+++ src/wai/annotations/format/tf/utils/__init__.py
@@ -1,6 +1,7 @@
 """
 Package for utilities for working with the Tensorflow Records format.
 """
+from ._dense_numerical_from_mask import dense_numerical_from_mask
 from ._extract_feature import extract_feature
 from ._ensure_available import tensorflow
 from ._format import TensorflowExampleExternalFormat
@@ -9,4 +10,5 @@ from ._LabelMapAccumulator import LabelMapAccumulator
 from ._make_feature import make_feature
 from ._mask_from_polygon import mask_from_polygon
 from ._negative_example import negative_example
+from ._png_from_mask import png_from_mask
 from ._polygon_from_mask import polygon_from_mask