+++ src/wai/annotations/format/tf/utils/_mask_from_polygon.py
@@ -1,13 +1,16 @@
-import io
+import numpy as np
 
-import PIL
 from wai.common.geometry import Polygon
 
 from ....core.util import polygon_to_poly_array
 from ....image_utils import polygon_to_mask
 
 
-def mask_from_polygon(polygon: Polygon, image_width: int, image_height: int) -> bytes:
+def mask_from_polygon(
+        polygon: Polygon,
+        image_width: int,
+        image_height: int
+) -> np.ndarray:
     """
     Creates a mask from an object's polygon. Based on code
     from:
@@ -17,16 +20,10 @@ def mask_from_polygon(polygon: Polygon, image_width: int, image_height: int) -> 
     :param polygon:         The object's bounding polygon.
     :param image_width:     The width of the image.
     :param image_height:    The height of the image.
-    :return:                The mask data.
+    :return:                An array covering the image, 1 inside the polygon and 0 outside.
     """
     # Create a Numpy array for the polygon
     poly_array = polygon_to_poly_array(polygon)
 
     # Run-length encode the polygon into a bitmask
-    binary_mask = polygon_to_mask([poly_array], image_width, image_height)
-
-    # Write the bitmask into PNG format
-    output_io = io.BytesIO()
-    PIL.Image.fromarray(binary_mask).save(output_io, format="PNG")
-
-    return output_io.getvalue()
+    return polygon_to_mask([poly_array], image_width, image_height)